+++ zerobuf/Vector.h
@@ -45,7 +45,8 @@ public:
     T* data() { return _alloc->template getDynamic< T >( _index ); }
 
     /** @return The pointer to the current allocation of the vector */
-    const T* data() const { return _alloc->template getDynamic< T >( _index ); }
+    const T* data() const
+        { return const_cast< const Allocator* >( _alloc )->template getDynamic< T >( _index ); }
 
     /** @return true if the two vectors of builtins are identical. */
     bool operator == ( const Vector& rhs ) const;